requirements.txt
requests
python-dotenv
googlemaps
langchain
langchain-ollama
fastapi
uvicorn[standard]
pydantic
